backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
motor==3.4.0
pymongo==4.6.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
authlib==1.3.0
httpx==0.27.0
anthropic==0.39.0
python-dotenv==1.0.0